+++ src/resources/views/widgets/filterset.blade.php
@@ -5,7 +5,7 @@
     @if($filters->activeCount() > 0)
         <span class="position-relative" title="{{ __('Filters') }}" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
             {!! icon('filters', null, ['class' => 'btn']) !!}
-            <small class="position-absolute" style="left: 25px; top: -7px"><span class="badge badge-danger">{{ $filters->activeCount() }}</span></small>
+            <small class="position-absolute" style="left: 25px; top: -7px"><span class="badge rounded-pill bg-danger">{{ $filters->activeCount() }}</span></small>
         </span>
     @else
         {!! icon('filters', null, [