+++ Changelog.md
@@ -6,6 +6,7 @@
 - Added Components based on the Laravel Blade Components Feature
 - Renamed the stylesheet of the default theme from `appshell` to `appshell3`
 - Changed the AppShell Theme - it has been rewritten from scratch with new visual style
+- Changed Breadcrumbs dependency from v7 to v8
 
 ## 3.1.2
 ##### 2022-08-27